Split void extension based on format
